moved storage ahead of browser tests
